Added XML file for device bind configurations
Updated the project to support an XML configuration file for defining device input bind settings. The new bindings.xml file contains definitions for both keyboard and gamepad controls. The project's packages versioning was also updated from fixed to dynamic and a method for acquiring input sources was added to the main game file.